Reduce measurement grid calculation from about 40ms to 4ms add methods for fusing polar grids

+++ include/opengl/polygon.h
@@ -24,7 +24,7 @@ public:
 class Polygon
 {
 public:
-	Polygon(Vertex* vertices, int num_vertices);
+	Polygon(Vertex* vertices, size_t num_vertices);
 	~Polygon();
 
 	void draw();
